docs(story-7-6-7): code review analysis complete with integer overflow fix [Story-7-6-7-error-report-cross-platform-diagnostics]
Documented BLOCKER finding and fix for integer overflow in RAM detection where
GetSystemInfo() assignments were casting to int before int64_t assignment.
Updated story artifacts (review, progress, story) with code review findings.

- The story shows "done" with all tasks checked. The pipeline feedback indicates the code-review-analysis found an incomplete integer overflow fix (`static_cast<int>` instead of `static_cast<int64_t>`). Let me verify the current code state.
- Quality gate background task output is empty — let me run it directly.
- Quality gate PASSED. All checks green. Now let me verify the story state and check the pipeline state file.
- `m_lpszGpuBackend` field in ErrorReport.h
- `WriteImeInfo(SDL_Window*)` signature

+++ src/source/Core/ErrorReport.h
@@ -13,7 +13,7 @@ typedef struct
 {
     wchar_t m_lpszCPU[MAX_LENGTH_CPUNAME];
     wchar_t m_lpszOS[MAX_LENGTH_OSINFO];
-    int m_iMemorySize;
+    int64_t m_iMemorySize; // int64_t to support all modern RAM sizes (>2GB)
 
     wchar_t m_lpszGpuBackend[MAX_DXVERSION];
 } ER_SystemInfo;
